baichun2-7B-chat-4bit/requirements/inference.txt
accelerate
colorama
bitsandbytes
sentencepiece
streamlit
transformers_stream_generator
cpm_kernels